Separation between base and http feature/dsl
manage the cornichon/core with a separation in transport concern
we could enable the use of TCP API as well.
