+++ cornichon-scalatest/src/main/scala/com/github/agourlay/cornichon/CornichonFeature.scala
@@ -1,6 +1,7 @@
 package com.github.agourlay.cornichon
 
-import com.github.agourlay.cornichon.feature.BaseFeature
+import com.github.agourlay.cornichon.feature.{ BaseFeature, HttpFeature }
 import com.github.agourlay.cornichon.scalatest.ScalatestFeature
 
-trait CornichonFeature extends BaseFeature with ScalatestFeature
+trait CornichonBaseFeature extends BaseFeature with ScalatestFeature
+trait CornichonFeature extends HttpFeature with ScalatestFeature
